Get by id en pedido detallePedido y producto

+++ src/main/java/com/tienda/tienda/repository/PedidoRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface PedidoRepo extends JpaRepository<Pedido, Long> {
-    Optional<Pedido> findPedidoById(int  id);
+    Optional<Pedido> findPedidoById(Long  id);
 
     void deletePedidoById(int id);
 }
